move sample, and fix splash screen positioning and rename blazor sample.
