+++ samples/ControlCatalog.Blazor.Web/Pages/Index.razor
