+++ samples/ControlCatalog.Blazor.Web/wwwroot/favicon.ico
