+++ samples/ControlCatalog.Blazor.Web/wwwroot/index.html
