+++ samples/ControlCatalog.Web/embed.js
