+++ samples/ControlCatalog.Web/favicon.ico
