+++ samples/ControlCatalog.Web/index.html
@@ -4,7 +4,7 @@
 <html>
 
 <head>
-    <title>Avalonia.Web.Sample</title>
+    <title>AvaloniaUI - ControlCatalog</title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="modulepreload" href="./main.js" />
