+++ samples/ControlCatalog.Web/main.js
